+++ src/rpc_api/mod.rs
@@ -80,15 +80,23 @@ pub static ACCESS_MAP: Lazy<HashMap<&str, Access>> = Lazy::new(|| {
     access.insert(state_api::STATE_GET_ACTOR, Access::Read);
     access.insert(state_api::STATE_MARKET_BALANCE, Access::Read);
     access.insert(state_api::STATE_MARKET_DEALS, Access::Read);
+    access.insert(state_api::STATE_MINER_INFO, Access::Read);
+    access.insert(state_api::STATE_MINER_ACTIVE_SECTORS, Access::Read);
+    access.insert(state_api::STATE_MINER_FAULTS, Access::Read);
     access.insert(state_api::STATE_MINER_POWER, Access::Read);
+    access.insert(state_api::STATE_MINER_DEADLINES, Access::Read);
     access.insert(state_api::STATE_GET_RECEIPT, Access::Read);
     access.insert(state_api::STATE_WAIT_MSG, Access::Read);
     access.insert(state_api::STATE_NETWORK_NAME, Access::Read);
     access.insert(state_api::STATE_NETWORK_VERSION, Access::Read);
+    access.insert(state_api::STATE_ACCOUNT_KEY, Access::Read);
+    access.insert(state_api::STATE_LOOKUP_ID, Access::Read);
     access.insert(state_api::STATE_FETCH_ROOT, Access::Read);
+    access.insert(state_api::STATE_GET_RANDOMNESS_FROM_TICKETS, Access::Read);
     access.insert(state_api::STATE_GET_RANDOMNESS_FROM_BEACON, Access::Read);
     access.insert(state_api::STATE_READ_STATE, Access::Read);
     access.insert(state_api::STATE_CIRCULATING_SUPPLY, Access::Read);
+    access.insert(state_api::STATE_SECTOR_GET_INFO, Access::Read);
 
     // Gas API
     access.insert(gas_api::GAS_ESTIMATE_GAS_LIMIT, Access::Read);
@@ -232,10 +240,14 @@ pub mod state_api {
     pub const STATE_GET_ACTOR: &str = "Filecoin.StateGetActor";
     pub const STATE_MARKET_BALANCE: &str = "Filecoin.StateMarketBalance";
     pub const STATE_MARKET_DEALS: &str = "Filecoin.StateMarketDeals";
+    pub const STATE_MINER_INFO: &str = "Filecoin.StateMinerInfo";
+    pub const STATE_MINER_FAULTS: &str = "Filecoin.StateMinerFaults";
     pub const STATE_MINER_POWER: &str = "Filecoin.StateMinerPower";
+    pub const STATE_MINER_DEADLINES: &str = "Filecoin.StateMinerDeadlines";
     pub const STATE_GET_RECEIPT: &str = "Filecoin.StateGetReceipt";
     pub const STATE_WAIT_MSG: &str = "Filecoin.StateWaitMsg";
     pub const STATE_FETCH_ROOT: &str = "Filecoin.StateFetchRoot";
+    pub const STATE_GET_RANDOMNESS_FROM_TICKETS: &str = "Filecoin.StateGetRandomnessFromTickets";
     pub const STATE_GET_RANDOMNESS_FROM_BEACON: &str = "Filecoin.StateGetRandomnessFromBeacon";
     pub const STATE_READ_STATE: &str = "Filecoin.StateReadState";
     pub const STATE_MINER_ACTIVE_SECTORS: &str = "Filecoin.StateMinerActiveSectors";
@@ -243,6 +255,9 @@ pub mod state_api {
     pub const STATE_ACCOUNT_KEY: &str = "Filecoin.StateAccountKey";
     pub const STATE_CIRCULATING_SUPPLY: &str = "Filecoin.StateCirculatingSupply";
     pub const STATE_DECODE_PARAMS: &str = "Filecoin.StateDecodeParams";
+    pub const STATE_SECTOR_GET_INFO: &str = "Filecoin.StateSectorGetInfo";
+    pub const STATE_SEARCH_MSG: &str = "Filecoin.StateSearchMsg";
+    pub const STATE_SEARCH_MSG_LIMITED: &str = "Filecoin.StateSearchMsgLimited";
 }
 
 /// Gas API