+++ src/rpc_client/state_ops.rs
@@ -6,16 +6,19 @@ use std::path::PathBuf;
 use crate::{
     blocks::TipsetKeys,
     rpc_api::{
-        data_types::{ApiActorState, SectorOnChainInfo},
+        data_types::{
+            ApiActorState, ApiDeadline, ApiInvocResult, MessageLookup, SectorOnChainInfo,
+        },
         state_api::*,
     },
     shim::{
-        address::Address, clock::ChainEpoch, econ::TokenAmount, message::MethodNum,
-        state_tree::ActorState, version::NetworkVersion,
+        address::Address, clock::ChainEpoch, econ::TokenAmount, message::Message,
+        message::MethodNum, state_tree::ActorState, version::NetworkVersion,
     },
 };
 use cid::Cid;
-use fil_actor_interface::miner::MinerPower;
+use fil_actor_interface::miner::{MinerInfo, MinerPower};
+use fil_actors_shared::fvm_ipld_bitfield::BitField;
 use fil_actors_shared::v10::runtime::DomainSeparationTag;
 use libipld_core::ipld::Ipld;
 
@@ -57,8 +60,39 @@ impl ApiInfo {
         RpcRequest::new(STATE_NETWORK_NAME, ())
     }
 
+    pub fn state_miner_info_req(miner: Address, tsk: TipsetKeys) -> RpcRequest<MinerInfo> {
+        RpcRequest::new(STATE_MINER_INFO, (miner, tsk))
+    }
+
+    pub fn state_call_req(message: Message, tsk: TipsetKeys) -> RpcRequest<ApiInvocResult> {
+        RpcRequest::new(STATE_CALL, (message, tsk))
+    }
+
+    pub fn state_miner_faults_req(miner: Address, tsk: TipsetKeys) -> RpcRequest<BitField> {
+        RpcRequest::new(STATE_MINER_FAULTS, (miner, tsk))
+    }
+
     pub fn state_miner_power_req(miner: Address, tsk: TipsetKeys) -> RpcRequest<MinerPower> {
         RpcRequest::new(STATE_MINER_POWER, (miner, tsk))
+    }
+
+    pub fn state_miner_deadlines_req(
+        miner: Address,
+        tsk: TipsetKeys,
+    ) -> RpcRequest<Vec<ApiDeadline>> {
+        RpcRequest::new(STATE_MINER_DEADLINES, (miner, tsk))
+    }
+
+    pub fn state_get_randomness_from_tickets_req(
+        tsk: TipsetKeys,
+        personalization: DomainSeparationTag,
+        rand_epoch: ChainEpoch,
+        entropy: Vec<u8>,
+    ) -> RpcRequest<Vec<u8>> {
+        RpcRequest::new(
+            STATE_GET_RANDOMNESS_FROM_TICKETS,
+            (personalization as i64, rand_epoch, entropy, tsk),
+        )
     }
 
     pub fn state_get_randomness_from_beacon_req(
@@ -107,5 +141,29 @@ impl ApiInfo {
         tsk: TipsetKeys,
     ) -> RpcRequest<Ipld> {
         RpcRequest::new(STATE_DECODE_PARAMS, (recipient, method_number, params, tsk))
+    }
+
+    pub fn state_sector_get_info_req(
+        addr: Address,
+        sector_no: u64,
+        tsk: TipsetKeys,
+    ) -> RpcRequest<SectorOnChainInfo> {
+        RpcRequest::new(STATE_SECTOR_GET_INFO, (addr, sector_no, tsk))
+    }
+
+    // FIXME: StateWaitMsg API gets stuck in forest
+    // pub fn state_wait_msg_req(msg_cid: Cid, confidence: i64) -> RpcRequest<Option<MessageLookup>> {
+    //     RpcRequest::new(STATE_WAIT_MSG, (msg_cid, confidence))
+    // }
+
+    pub fn state_search_msg_req(msg_cid: Cid) -> RpcRequest<Option<MessageLookup>> {
+        RpcRequest::new(STATE_SEARCH_MSG, (msg_cid,))
+    }
+
+    pub fn state_search_msg_limited_req(
+        msg_cid: Cid,
+        limit_epoch: i64,
+    ) -> RpcRequest<Option<MessageLookup>> {
+        RpcRequest::new(STATE_SEARCH_MSG_LIMITED, (msg_cid, limit_epoch))
     }
 }